==> empirical/data_mining_gta/9 conditional.py <==
# -*-coding: utf-8 -*-
# Python 3.6
# Author:Zhang Haitao
# Email:13163385579@163.com
# TIME:2018-09-19  09:08
# NAME:FT_hp-9 conditional.py

import pandas as pd
from config import DIR_TMP
import os

from empirical.config_ep import DIR_DM_GTA, CROSS_LEAST
from empirical.data_mining_gta.analyze_playing_field8 import \
    get_prominent_indicators
from empirical.get_basedata import BENCHS
from tools import multi_process

# CONDITIONAL='ivol'

DIR_COND=os.path.join(DIR_DM_GTA,'analyse','conditional')

#conditional on idiosyncratical volatility
def get_indicator_direction(indname):
    at = pd.concat(
        [pd.read_pickle(os.path.join(DIR_DM_GTA,'analyse', f'at_{bench}.pkl'))
         for bench in BENCHS], axis=1,sort=True,keys=BENCHS)

    symbol=1 if at.at[indname,'capmM']>0 else -1
    return symbol


def get_comb_indicators(critic=2):
    iid='oepwqfaldagha'
    path=os.path.join(DIR_COND,f'{iid}{critic}.pkl')
    # path=os.path.join(DIR_TMP,'dafj3e3ncsv.pkl')
    if os.path.exists(path):
        return pd.read_pickle(path)
    else:
        ind_names = get_prominent_indicators(critic)
        indicators = pd.concat([pd.read_pickle(
            os.path.join(DIR_DM_GTA,'normalized',
                         ind + '.pkl')).stack() * get_indicator_direction(ind) for
                                ind in
                                ind_names], axis=1,
                               keys=ind_names)  # trick：revert the negative signal
        indicators.to_pickle(path)
        return indicators

def get_comb(cond_variable):
    indicators=get_comb_indicators()
    conditional=pd.read_pickle(os.path.join(DIR_DM_GTA,'conditional',cond_variable+'.pkl')).shift(1).stack()
    conditional.name=cond_variable
    # conditional=pd.read_pickle(os.path.join(DIR_BASEDATA,'normalized_conditional',cond_variable+'.pkl'))
    # ret=get_filtered_ret().swaplevel()
    ret=pd.read_pickle(os.path.join(DIR_DM_GTA,'fdmt_m.pkl'))['ret_m']
    # conditional=conditional.groupby('stkcd').shift(1)
    indicators=indicators.groupby('stkcd').shift(1)#trick: use the indicators of time t

    comb=pd.concat([indicators, ret, conditional], axis=1)
    comb.index.names=['month_end','stkcd']

    comb=comb.dropna(subset=['ret_m',cond_variable])
    comb=comb.groupby('month_end').filter(lambda df:df.shape[0]>CROSS_LEAST)#trick: filter out months with too small sample
    comb=comb.fillna(0)
    print(cond_variable,len(comb.index.get_level_values('month_end').unique()))
    return comb


def _conditional_anomaly_return(args):
    comb,col,cond_variable=args
    #groupby conditional variable
    comb['gc']=comb.groupby('month_end',group_keys=False).apply(lambda df:pd.qcut(df[cond_variable].rank(method='first'),5,
                                                          labels=[f'g{i}' for i in range(1,6)]))

    #groupby factor
    comb['gf']=comb.groupby(['month_end','gc'],group_keys=False).apply(lambda df:pd.qcut(df[col].rank(method='first'),10,
                                labels=[f'g{i}' for i in range(1,11)]))

    stk=comb.groupby(['month_end', 'gc', 'gf']).apply(lambda df:df['ret_m'].mean()).unstack('gf')
    panel=(stk['g10'] - stk['g1']).unstack()
    panel.columns=panel.columns.astype(str)
    panel['all']=panel.mean(axis=1)
    panel['high-low']=panel['g5']-panel['g1']

    alpha=panel.mean()
    # t=panel.mean()/panel.std()
    t=panel.mean()/panel.sem() #trick: tvalue = mean / stderr,   stderr = std / sqrt(n-1) ,pd.Series.sem() = pd.Series.std()/pow(len(series),0.5)

    table=pd.concat([alpha,t],axis=1,keys=['alpha','t']).T
    # print(col)
    return table

def get_conditional_anomaly_return(cond_variable):
    comb = get_comb(cond_variable)
    indicators = [col for col in comb.columns if
                  col not in ['ret_m', cond_variable]]

    args_generator=((comb,col,cond_variable) for col in indicators)

    # tables=[]
    # for args in args_generator:
    #     print(args[1])
    #     tables.append(_conditional_anomaly_return(args))#fixme:
    tables=multi_process(_conditional_anomaly_return,args_generator,20)
    table5=pd.concat(tables,axis=0,keys=indicators)
    table5.to_csv(os.path.join(DIR_COND, f'table5_{cond_variable}.csv'))



def test_all_conditional_indicators():
    fns = os.listdir(os.path.join(DIR_DM_GTA, 'conditional'))
    conds = [fn[:-4] for fn in fns]
    for cond in conds[7:]:#fixme:
        get_conditional_anomaly_return(cond)
        print(cond)

def debug():
    cond='turnover_120'
    get_comb(cond)
    # get_conditional_anomaly_return(cond)



def analyze_turnover():
    fns=os.listdir(os.path.join(DIR_DM_GTA,'analyse','conditional'))
    fns=[fn for fn in fns if fn.startswith('table5_turnover')]
    dfs = [pd.read_csv(os.path.join(DIR_DM_GTA,'analyse','conditional', fn), index_col=0) for fn in
           fns]
    df = pd.concat(dfs, keys=[fn[:-4] for fn in fns]) #fixme:
    df = df[df.iloc[:, 0] == 't']
    df['abs'] = df['high-low'].abs()
    df = df.sort_values('abs', ascending=False)
    df = df.iloc[:, 1:]
    df.index.names = ['cond_variable', 'indicator']

    lt2 = df.groupby('cond_variable').apply(
        lambda df: df[df['abs'] >= 2].shape[0]).sort_values(ascending=False)
    target = df.loc[(lt2.index[0], slice(None)), :]
    target.to_csv(os.path.join(DIR_DM_GTA,'analyse','conditional','turnover.csv'))


def analyze_idio():
    fns=os.listdir(os.path.join(DIR_DM_GTA,'analyse','conditional'))
    fns=[fn for fn in fns if fn.startswith('table5_idio')]
    dfs = [pd.read_csv(os.path.join(DIR_DM_GTA,'analyse','conditional', fn), index_col=0) for fn in
           fns]
    df = pd.concat(dfs, keys=[fn[:-4] for fn in fns]) #fixme:
    df = df[df.iloc[:, 0] == 't']
    df['abs'] = df['high-low'].abs()
    df = df.sort_values('abs', ascending=False)
    df = df.iloc[:, 1:]
    df.index.names = ['cond_variable', 'indicator']

    lt2 = df.groupby('cond_variable').apply(
        lambda df: df[df['abs'] >= 2].shape[0]).sort_values(ascending=False)
    target = df.loc[(lt2.index[0], slice(None)), :]
    target.to_csv(os.path.join(DIR_DM_GTA,'analyse','conditional','idio.csv'))


if __name__ == '__main__':
    analyze_idio()


def main():
    test_all_conditional_indicators()


# if __name__ == '__main__':
#     main()
